Overhaul naming scheme to be more like C stdlib, other function changes.
oystr_write was removed, due to not being super useful. oystr_ensure_size was
renamed to stxensure_size. stx_grow is now used by stx functions that manage
memory rather than stxensure_size, as the stx->mem member shouldn't be allocated
in a doubling fashion, as that can waste a lot of memory. Instead, the
responsiblity of doubling memory is left to the user of the library, as they
must now call stxensure_size when they want to grow the buffer as exponents of
2. Tests have regressed a little, as they need to be rewritten to support the
new naming scheme.