Modify return values to allow for function composition
Most functions that one would chain together have been changed to return a
'struct stx*' rather than an 'int'. This was made to allow for easy function
composition, and to facilicate this change, the error return values have been
simplified to these functions now simply return NULL if anything has gone wrong.

Note that functions still do not validate their parameters, so the safest way
to use this library is to still check the return value line-by-line. This change
was to make it easier to decouple the growth functions from the rest of the stx
functions.